Building fundamental classes & GUI

+++ src/main/java/com/example/merchlogistics/Entities/Calculable.java
@@ -1,5 +1,5 @@
 package com.example.merchlogistics.Entities;
 
 public interface Calculable {
-    double calculate();
+    int calculate();
 }
